backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
alembic>=1.13.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
email-validator==2.1.0
python-dotenv>=1.0.0
httpx>=0.26.0
starlette==0.35.1
itsdangerous==2.1.2  # Required for session middleware
psycopg2-binary>=2.9.9
requests>=2.31.0
aiohttp>=3.9.0
openai>=1.12.0  # OpenAI API client
anthropic>=0.18.0  # Anthropic API client
transformers>=4.38.0
torch>=2.2.0
accelerate>=0.27.0
sentence-transformers==2.5.1
numpy==1.26.4
pandas==2.2.0
scikit-learn==1.6.1
matplotlib==3.8.3
seaborn==0.13.2
aiofiles==23.2.1  # For file operations
cachetools==5.3.2  # For in-memory caching
diskcache==5.6.3  # For filesystem caching
APScheduler==3.10.4  # For task scheduling

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov==4.1.0

# Code Quality
black==24.1.1
isort==5.13.2
mypy==1.8.0
ruff==0.2.1

# ML/AI
torchvision==0.15.2

# Visualization
matplotlib==3.8.3
seaborn==0.13.2

# Development
aiofiles==23.2.1  # For file operations
cachetools==5.3.2  # For in-memory caching
diskcache==5.6.3  # For filesystem caching
APScheduler==3.10.4  # For task scheduling

# Additional dependencies
tenacity>=8.2.0
redis>=5.0.0
boto3>=1.34.0
sentry-sdk>=1.40.0
prometheus-client>=0.19.0 